requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
tensorflow==2.16.1  # if you're using TensorFlow
# torch==2.2.2        # Uncomment this if you're using PyTorch instead
Pillow==10.3.0
nltk==3.8.1
textblob==0.18.0
firebase-admin==6.5.0  # only if using Firebase
joblib==1.4.2
plotly==5.21.0
